finish almost students account tabs

+++ app/Repositories/Repository/FeeProcessingRepository.php
@@ -27,38 +27,35 @@ class FeeProcessingRepository implements FeeProcessingRepositoryInterface
 
                 $data               = $request->only(['student_id', 'amount', 'student_funds', 'description',]);
                 $studentFunds       = StudentAccount::where('student_id', $data['student_id'])->sum('debit') - StudentAccount::where('student_id', $data['student_id'])->sum('credit');
-                if ($studentFunds   == $data['student_funds']) {
-                    if ($data['amount'] > $studentFunds && $studentFunds != 0) {
-                        toastr()->error(__('msgs.amount_warning', ['amount' => $studentFunds]));
+
+
+
+                if ($data['amount'] > $studentFunds && $studentFunds != 0) {
+                    toastr()->error(__('msgs.amount_warning', ['amount' => $studentFunds]));
+                    return redirect()->back();
+                } else {
+                    if ($studentFunds <= 0) {
+                        toastr()->error(__('msgs.student_funds_is_zero'));
                         return redirect()->back();
                     } else {
-                        if ($studentFunds <= 0) {
-                            toastr()->error(__('msgs.student_funds_is_zero'));
-                            return redirect()->back();
-                        } else {
-                            $feeProcessing      = FeeProcessing::create([
-                                'student_id'        => $data['student_id'],
-                                'amount'            => $data['amount'],
-                                'description'       => $data['description']
-                            ]);
+                        $feeProcessing      = FeeProcessing::create([
+                            'student_id'        => $data['student_id'],
+                            'amount'            => $data['amount'],
+                            'description'       => $data['description']
+                        ]);
 
-                            StudentAccount::create([
-                                'type'              => 'exclusion',
-                                'student_id'        => $data['student_id'],
-                                'feeProcessing_id'  => FeeProcessing::latest()->first()->id,
-                                'credit'            => $data['amount'],
-                                'description'       => $data['description'],
+                        StudentAccount::create([
+                            'type'              => 'exclusion',
+                            'student_id'        => $data['student_id'],
+                            'feeProcessing_id'  => FeeProcessing::latest()->first()->id,
+                            'credit'            => $data['amount'],
+                        ]);
 
-                            ]);
+                        DB::commit();
 
-                            DB::commit();
-
-                            toastr()->success(__('msgs.added', ['name' => __('fee.fee_exclusion')]));
-                            return redirect()->back();
-                        }
+                        toastr()->success(__('msgs.added', ['name' => __('fee.fee_exclusion')]));
+                        return redirect()->back();
                     }
-                } else {
-                    return redirect()->back();
                 }
             }
         } catch (\Throwable $th) {
@@ -69,17 +66,12 @@ class FeeProcessingRepository implements FeeProcessingRepositoryInterface
 
     public function edit($feeProcessing)
     {
-        $student                = Student::findOrFail($feeProcessing->student_id);
-        $studentReceipts        = StudentReceipt::where('student_id', $feeProcessing->student_id)->get();
-        $feeProcessings         = FeeProcessing::where('student_id', $feeProcessing->student_id)->get();
-        $studentFeeInvoices     = FeeInvoice::with(['student', 'fee'])->where('student_id', $feeProcessing->student_id)->get();
-        return view('pages.fee-processings.edit', [
-            'feeProcessing'         => $feeProcessing,
-            'student'               => $student,
-            'studentReceipts'       => $studentReceipts,
-            'studentFeeInvoices'    => $studentFeeInvoices,
-            'feeProcessings'        => $feeProcessings
-        ]);
+        $data['feeProcessing']          = $feeProcessing;
+        $data['student']                = Student::findOrFail($feeProcessing->student_id);
+        $data['studentReceipts']        = StudentReceipt::where('student_id', $feeProcessing->student_id)->get();
+        $data['feeProcessings']         = FeeProcessing::where('student_id', $feeProcessing->student_id)->get();
+        $data['studentFeeInvoices']     = FeeInvoice::with(['student', 'fee'])->where('student_id', $feeProcessing->student_id)->get();
+        return view('pages.fee-processings.edit', $data);
     }
 
     public function update($request, $feeProcessing)
@@ -89,7 +81,10 @@ class FeeProcessingRepository implements FeeProcessingRepositoryInterface
             if ($request->isMethod('put')) {
                 $data               = $request->only(['student_id', 'amount', 'student_funds', 'description',]);
                 $studentFunds       = StudentAccount::where('student_id', $data['student_id'])->sum('debit') - StudentAccount::where('student_id', $data['student_id'])->sum('credit');
-                if ($studentFunds   == $data['student_funds']) {
+
+                if ($data['amount'] > $studentFunds && $studentFunds != 0) {
+                    toastr()->error(__('msgs.amount_warning', ['amount' => $studentFunds]));
+                } else {
                     $feeProcessing->update([
                         'amount'            => $data['amount'],
                         'description'       => $data['description']
@@ -97,16 +92,13 @@ class FeeProcessingRepository implements FeeProcessingRepositoryInterface
 
                     StudentAccount::where('student_id', $data['student_id'])->where('feeProcessing_id', $feeProcessing->id)->update([
                         'credit'            => $data['amount'],
-                        'description'       => $data['description'],
                     ]);
 
                     DB::commit();
-
                     toastr()->success(__('msgs.updated', ['name' => __('fee.fee_exclusion')]));
-                    return redirect()->back();
-                } else {
-                    return redirect()->back();
                 }
+
+                return redirect()->back();
             }
         } catch (\Throwable $th) {
             DB::rollBack();
@@ -117,8 +109,15 @@ class FeeProcessingRepository implements FeeProcessingRepositoryInterface
     public function destroy($feeProcessing)
     {
         try {
-            $feeProcessing->delete();
-            toastr()->info(__('msgs.deleted', ['name' => __('fee.fee_exclusion')]));
+            $student    = Student::where('id', $feeProcessing->student_id)->first();
+
+            if ($student)
+                toastr()->error(__('msgs.is_existed', ['name' => $student->name]));
+            else {
+                $feeProcessing->delete();
+                toastr()->info(__('msgs.deleted', ['name' => __('fee.fee_exclusion')]));
+            }
+
             return redirect()->back();
         } catch (\Throwable $th) {
             return redirect()->back()->withErrors(['error' => $th->getMessage()]);
@@ -127,15 +126,10 @@ class FeeProcessingRepository implements FeeProcessingRepositoryInterface
 
     public function addFeeExclusion($student_id)
     {
-        $student                = Student::findOrFail($student_id);
-        $studentReceipts        = StudentReceipt::where('student_id', $student_id)->get();
-        $feeProcessings         = FeeProcessing::where('student_id', $student_id)->get();
-        $studentFeeInvoices     = FeeInvoice::with(['student', 'fee'])->where('student_id', $student_id)->get();
-        return view('pages.fee-processings.create', [
-            'student'               => $student,
-            'studentReceipts'       => $studentReceipts,
-            'studentFeeInvoices'    => $studentFeeInvoices,
-            'feeProcessings'        => $feeProcessings
-        ]);
+        $data['student']                = Student::findOrFail($student_id);
+        $data['studentReceipts']        = StudentReceipt::where('student_id', $student_id)->get();
+        $data['feeProcessings']         = FeeProcessing::where('student_id', $student_id)->get();
+        $data['studentFeeInvoices']     = FeeInvoice::with(['student', 'fee'])->where('student_id', $student_id)->get();
+        return view('pages.fee-processings.create', $data);
     }
 }
