+++ app/Repositories/Repository/StudentPaymentRepository.php
@@ -22,18 +22,12 @@ class StudentPaymentRepository implements StudentPaymentRepositoryInterface
 
     public function addStudentPayment($student_id)
     {
-        $student                = Student::findOrFail($student_id);
-        $studentFeeInvoices     = FeeInvoice::with(['student', 'fee'])->where('student_id', $student_id)->get();
-        $studentReceipts        = StudentReceipt::where('student_id', $student_id)->get();
-        $feeProcessings         = FeeProcessing::where('student_id', $student_id)->get();
-        $studentPayments        = StudentPayment::where('student_id', $student_id)->get();
-        return view('pages.student-payments.create', [
-            'student'               => $student,
-            'studentFeeInvoices'    => $studentFeeInvoices,
-            'studentReceipts'       => $studentReceipts,
-            'feeProcessings'        => $feeProcessings,
-            'studentPayments'       => $studentPayments
-        ]);
+        $data['student']                = Student::findOrFail($student_id);
+        $data['studentFeeInvoices']     = FeeInvoice::with(['student', 'fee'])->where('student_id', $student_id)->get();
+        $data['studentReceipts']        = StudentReceipt::where('student_id', $student_id)->get();
+        $data['feeProcessings']         = FeeProcessing::where('student_id', $student_id)->get();
+        $data['studentPayments']        = StudentPayment::where('student_id', $student_id)->get();
+        return view('pages.student-payments.create', $data);
     }
 
 
@@ -79,19 +73,13 @@ class StudentPaymentRepository implements StudentPaymentRepositoryInterface
 
     public function edit($studentPayment)
     {
-        $student                = Student::findOrFail($studentPayment->student_id);
-        $studentFeeInvoices     = FeeInvoice::with(['student', 'fee'])->where('student_id', $studentPayment->student_id)->get();
-        $studentReceipts        = StudentReceipt::where('student_id', $studentPayment->student_id)->get();
-        $feeProcessings         = FeeProcessing::where('student_id', $studentPayment->student_id)->get();
-        $studentPayments        = StudentPayment::where('student_id', $studentPayment->student_id)->get();
-        return view('pages.student-payments.edit', [
-            'studentPayment'        => $studentPayment,
-            'student'               => $student,
-            'studentFeeInvoices'    => $studentFeeInvoices,
-            'studentReceipts'       => $studentReceipts,
-            'feeProcessings'        => $feeProcessings,
-            'studentPayments'       => $studentPayments,
-        ]);
+        $data['studentPayment']         = $studentPayment;
+        $data['student']                = Student::findOrFail($studentPayment->student_id);
+        $data['studentFeeInvoices']     = FeeInvoice::with(['student', 'fee'])->where('student_id', $studentPayment->student_id)->get();
+        $data['studentReceipts']        = StudentReceipt::where('student_id', $studentPayment->student_id)->get();
+        $data['feeProcessings']         = FeeProcessing::where('student_id', $studentPayment->student_id)->get();
+        $data['studentPayments']        = StudentPayment::where('student_id', $studentPayment->student_id)->get();
+        return view('pages.student-payments.edit',);
     }
 
     public function update($request, $studentPayment)
@@ -130,8 +118,15 @@ class StudentPaymentRepository implements StudentPaymentRepositoryInterface
     public function destroy($studentPayment)
     {
         try {
-            $studentPayment->delete();
-            toastr()->info(__('msgs.deleted', ['name' => __('fee.payment')]));
+            $student    = Student::where('id', $studentPayment->student_id)->first();
+
+            if ($student)
+                toastr()->error(__('msgs.is_existed', ['name' => $student->name]));
+            else {
+                $studentPayment->delete();
+                toastr()->info(__('msgs.deleted', ['name' => __('fee.payment')]));
+            }
+
             return redirect()->back();
         } catch (\Throwable $th) {
             return redirect()->back()->withErrors(['error' => $th->getMessage()]);
